requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
anyio==4.12.0
asyncmy==0.2.10
atlastk==0.13.5
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
discord-webhook==1.4.1
disnake==2.11.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
PyMySQL==1.1.2
pypika-tortoise==0.6.3
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
tortoise-orm==0.25.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
cachetools>=5.3.0
